KOGITO-1681: Moving TestRunnerReportingPanelWrapper & Impl in BC module only.
